Projects/myToothless/requirements.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
geojson==2.5.0
google-search-results==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.99
llvmlite==0.43.0
MarkupSafe==2.1.5
marshmallow==3.21.3
more-itertools==10.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai==1.41.0
openai-whisper==20231117
orjson==3.10.7
packaging==24.1
pvporcupine==3.0.2
pvrecorder==1.2.2
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.20.1
pyowm==3.3.0
PySocks==1.7.1
python-dotenv==1.0.1
python-vlc==3.0.20123
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.32
sympy==1.13.2
tenacity==8.5.0
tiktoken==0.7.0
torch==2.4.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
